web-app/requirements.txt
flask
flask-bootstrap
flask-bcrypt
flask-sqlalchemy
flask-wtf
flask-mail
flask-login
Flask-Migrate
wtforms
PyJWT
flask-blueprint
python-dotenv
uWSGI
mysqlclient